refactor: simplify work with modal visibility
